monitor/app/requirements.txt
boto3
pynacl
mcipc